requirements.txt
langchain==0.3.19
langchain_community==0.3.18
langchain_openai==0.3.7
PyMuPDF==1.22.2
python-dotenv==1.0.0
selenium==4.16.0
sentence_transformers==3.4.1
